issue124: extract Difference
